programs/ian-agency/contexts/samlino/seo-agent-playground/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.48.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.1
python-multipart>=0.0.6
